+++ spec/rubocop/cop/style/single_line_methods_spec.rb
@@ -28,6 +28,12 @@ describe RuboCop::Cop::Style::SingleLineMethods do
                            'def @table.columns; end'])
       expect(cop.offenses.size).to eq(3)
     end
+
+    it 'auto-corrects an empty method' do
+      corrected = autocorrect_source(cop, 'def x; end')
+      expect(corrected).to eq(['def x; ',
+                               'end'].join("\n"))
+    end
   end
 
   context 'when AllowIfMethodIsEmpty is enabled' do